Updated HTML and SASS
In this commit, added the whole html structure of the page and style. Added as well some basic JS for mobile responsiveness and popup window and navigation active link. Added as well additional folder document for my resume file and added as well all the necessary and temporary images and logo.